+++ Configuration/TCA/Overrides/sys_template.php
@@ -12,4 +12,5 @@ defined('TYPO3') || die();
 // add static typoscript includes
 \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addStaticFile('content_animations', 'Configuration/TypoScript', 'Content Animations: Basic Configuration');
 \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addStaticFile('content_animations', 'Configuration/TypoScript/BootstrapPackage/v15', 'Content Animations: Bootstrap Package v15.x');
+\TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addStaticFile('content_animations', 'Configuration/TypoScript/ContentBlocks', 'Content Animations: Content Blocks');
 \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addStaticFile('content_animations', 'Configuration/TypoScript/FluidStyledContent', 'Content Animations: Fluid Styled Content');