+++ Configuration/TCA/Overrides/tt_content.php
@@ -18,38 +18,161 @@ defined('TYPO3') || die();
             'label' => 'LLL:EXT:content_animations/Resources/Private/Language/locallang_be.xlf:label.animation',
             'config' => [
                 'items' => [
-                    ['LLL:EXT:content_animations/Resources/Private/Language/locallang_be.xlf:item.no-animation', ''],
-                    ['LLL:EXT:content_animations/Resources/Private/Language/locallang_be.xlf:item.fade-animations', '--div--'],
-                    ['fade', 'fade', 'EXT:content_animations/Resources/Public/Images/fade.gif'],
-                    ['fade-up', 'fade-up', 'EXT:content_animations/Resources/Public/Images/fade-up.gif'],
-                    ['fade-down', 'fade-down', 'EXT:content_animations/Resources/Public/Images/fade-down.gif'],
-                    ['fade-right', 'fade-right', 'EXT:content_animations/Resources/Public/Images/fade-right.gif'],
-                    ['fade-left', 'fade-left', 'EXT:content_animations/Resources/Public/Images/fade-left.gif'],
-                    ['fade-up-right', 'fade-up-right', 'EXT:content_animations/Resources/Public/Images/fade-up-right.gif'],
-                    ['fade-up-left', 'fade-up-left', 'EXT:content_animations/Resources/Public/Images/fade-up-left.gif'],
-                    ['fade-down-right', 'fade-down-right', 'EXT:content_animations/Resources/Public/Images/fade-down-right.gif'],
-                    ['fade-down-left', 'fade-down-left', 'EXT:content_animations/Resources/Public/Images/fade-down-left.gif'],
-                    ['LLL:EXT:content_animations/Resources/Private/Language/locallang_be.xlf:item.flip-animations', '--div--'],
-                    ['flip-up', 'flip-up', 'EXT:content_animations/Resources/Public/Images/flip-up.gif'],
-                    ['flip-down', 'flip-down', 'EXT:content_animations/Resources/Public/Images/flip-down.gif'],
-                    ['flip-left', 'flip-left', 'EXT:content_animations/Resources/Public/Images/flip-left.gif'],
-                    ['flip-right', 'flip-right', 'EXT:content_animations/Resources/Public/Images/flip-right.gif'],
-                    ['LLL:EXT:content_animations/Resources/Private/Language/locallang_be.xlf:item.slide-animations', '--div--'],
-                    ['slide-up', 'slide-up', 'EXT:content_animations/Resources/Public/Images/slide-up.gif'],
-                    ['slide-down', 'slide-down', 'EXT:content_animations/Resources/Public/Images/slide-down.gif'],
-                    ['slide-right', 'slide-right', 'EXT:content_animations/Resources/Public/Images/slide-right.gif'],
-                    ['slide-left', 'slide-left', 'EXT:content_animations/Resources/Public/Images/slide-left.gif'],
-                    ['LLL:EXT:content_animations/Resources/Private/Language/locallang_be.xlf:item.zoom-animations', '--div--'],
-                    ['zoom-in', 'zoom-in', 'EXT:content_animations/Resources/Public/Images/zoom-in.gif'],
-                    ['zoom-in-up', 'zoom-in-up', 'EXT:content_animations/Resources/Public/Images/zoom-in-up.gif'],
-                    ['zoom-in-down', 'zoom-in-down', 'EXT:content_animations/Resources/Public/Images/zoom-in-down.gif'],
-                    ['zoom-in-right', 'zoom-in-right', 'EXT:content_animations/Resources/Public/Images/zoom-in-right.gif'],
-                    ['zoom-in-left', 'zoom-in-left', 'EXT:content_animations/Resources/Public/Images/zoom-in-left.gif'],
-                    ['zoom-out', 'zoom-out', 'EXT:content_animations/Resources/Public/Images/zoom-out.gif'],
-                    ['zoom-out-up', 'zoom-out-up', 'EXT:content_animations/Resources/Public/Images/zoom-out-up.gif'],
-                    ['zoom-out-down', 'zoom-out-down', 'EXT:content_animations/Resources/Public/Images/zoom-out-down.gif'],
-                    ['zoom-out-right', 'zoom-out-right', 'EXT:content_animations/Resources/Public/Images/zoom-out-right.gif'],
-                    ['zoom-out-left', 'zoom-out-left', 'EXT:content_animations/Resources/Public/Images/zoom-out-left.gif'],
+                    [
+                        'label' => 'LLL:EXT:content_animations/Resources/Private/Language/locallang_be.xlf:item.no-animation',
+                        'value' => '',
+                    ],
+                    [
+                        'label' => 'LLL:EXT:content_animations/Resources/Private/Language/locallang_be.xlf:item.fade-animations',
+                        'value' => '--div--',
+                    ],
+                    [
+                        'label' => 'fade',
+                        'value' => 'fade',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/fade.gif',
+                    ],
+                    [
+                        'label' => 'fade-up',
+                        'value' => 'fade-up',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/fade-up.gif',
+                    ],
+                    [
+                        'label' => 'fade-down',
+                        'value' => 'fade-down',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/fade-down.gif',
+                    ],
+                    [
+                        'label' => 'fade-right',
+                        'value' => 'fade-right',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/fade-right.gif',
+                    ],
+                    [
+                        'label' => 'fade-left',
+                        'value' => 'fade-left',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/fade-left.gif',
+                    ],
+                    [
+                        'label' => 'fade-up-right',
+                        'value' => 'fade-up-right',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/fade-up-right.gif',
+                    ],
+                    [
+                        'label' => 'fade-up-left',
+                        'value' => 'fade-up-left',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/fade-up-left.gif',
+                    ],
+                    [
+                        'label' => 'fade-down-right',
+                        'value' => 'fade-down-right',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/fade-down-right.gif',
+                    ],
+                    [
+                        'label' => 'fade-down-left',
+                        'value' => 'fade-down-left',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/fade-down-left.gif',
+                    ],
+                    [
+                        'label' => 'LLL:EXT:content_animations/Resources/Private/Language/locallang_be.xlf:item.flip-animations',
+                        'value' => '--div--',
+                    ],
+                    [
+                        'label' => 'flip-up',
+                        'value' => 'flip-up',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/flip-up.gif',
+                    ],
+                    [
+                        'label' => 'flip-down',
+                        'value' => 'flip-down',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/flip-down.gif',
+                    ],
+                    [
+                        'label' => 'flip-left',
+                        'value' => 'flip-left',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/flip-left.gif',
+                    ],
+                    [
+                        'label' => 'flip-right',
+                        'value' => 'flip-right',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/flip-right.gif',
+                    ],
+                    [
+                        'label' => 'LLL:EXT:content_animations/Resources/Private/Language/locallang_be.xlf:item.slide-animations',
+                        'value' => '--div--',
+                    ],
+                    [
+                        'label' => 'slide-up',
+                        'value' => 'slide-up',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/slide-up.gif',
+                    ],
+                    [
+                        'label' => 'slide-down',
+                        'value' => 'slide-down',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/slide-down.gif',
+                    ],
+                    [
+                        'label' => 'slide-right',
+                        'value' => 'slide-right',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/slide-right.gif',
+                    ],
+                    [
+                        'label' => 'slide-left',
+                        'value' => 'slide-left',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/slide-left.gif',
+                    ],
+                    [
+                        'label' => 'LLL:EXT:content_animations/Resources/Private/Language/locallang_be.xlf:item.zoom-animations',
+                        'value' => '--div--',
+                    ],
+                    [
+                        'label' => 'zoom-in',
+                        'value' => 'zoom-in',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/zoom-in.gif',
+                    ],
+                    [
+                        'label' => 'zoom-in-up',
+                        'value' => 'zoom-in-up',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/zoom-in-up.gif',
+                    ],
+                    [
+                        'label' => 'zoom-in-down',
+                        'value' => 'zoom-in-down',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/zoom-in-down.gif',
+                    ],
+                    [
+                        'label' => 'zoom-in-right',
+                        'value' => 'zoom-in-right',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/zoom-in-right.gif',
+                    ],
+                    [
+                        'label' => 'zoom-in-left',
+                        'value' => 'zoom-in-left',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/zoom-in-left.gif',
+                    ],
+                    [
+                        'label' => 'zoom-out',
+                        'value' => 'zoom-out',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/zoom-out.gif',
+                    ],
+                    [
+                        'label' => 'zoom-out-up',
+                        'value' => 'zoom-out-up',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/zoom-out-up.gif',
+                    ],
+                    [
+                        'label' => 'zoom-out-down',
+                        'value' => 'zoom-out-down',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/zoom-out-down.gif',
+                    ],
+                    [
+                        'label' => 'zoom-out-right',
+                        'value' => 'zoom-out-right',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/zoom-out-right.gif',
+                    ],
+                    [
+                        'label' => 'zoom-out-left',
+                        'value' => 'zoom-out-left',
+                        'icon' => 'EXT:content_animations/Resources/Public/Images/zoom-out-left.gif',
+                    ],
                 ],
                 'renderType' => 'animationPreview',
                 'type' => 'select',
@@ -60,9 +183,8 @@ defined('TYPO3') || die();
             'exclude' => true,
             'label' => 'LLL:EXT:content_animations/Resources/Private/Language/locallang_be.xlf:label.duration',
             'config' => [
-                'type' => 'input',
+                'type' => 'number',
                 'size' => 5,
-                'eval' => 'trim,int',
                 'range' => [
                     'lower' => 400,
                     'upper' => 3000,
@@ -78,9 +200,8 @@ defined('TYPO3') || die();
             'exclude' => true,
             'label' => 'LLL:EXT:content_animations/Resources/Private/Language/locallang_be.xlf:label.delay',
             'config' => [
-                'type' => 'input',
+                'type' => 'number',
                 'size' => 5,
-                'eval' => 'trim,int',
                 'range' => [
                     'lower' => 0,
                     'upper' => 3000,
@@ -92,36 +213,62 @@ defined('TYPO3') || die();
                 ],
             ],
         ],
-
         'tx_content_animations_offset' => [
             'exclude' => true,
             'label' => 'LLL:EXT:content_animations/Resources/Private/Language/locallang_be.xlf:label.offset',
             'config' => [
-                'type' => 'input',
+                'type' => 'number',
                 'size' => 5,
-                'eval' => 'trim,int',
                 'default' => 0,
             ],
         ],
-
         'tx_content_animations_anchor_placement' => [
             'exclude' => true,
             'label' => 'LLL:EXT:content_animations/Resources/Private/Language/locallang_be.xlf:label.anchor-placement',
             'config' => [
                 'type' => 'select',
                 'renderType' => 'selectSingle',
-                //'size' => 1,
                 'items' => [
-                    ['', ''],
-                    ['top-bottom', 'top-bottom'],
-                    ['top-center', 'top-center'],
-                    ['top-top', 'top-top'],
-                    ['center-bottom', 'center-bottom'],
-                    ['center-center', 'center-center'],
-                    ['center-top', 'center-top'],
-                    ['bottom-bottom', 'bottom-bottom'],
-                    ['bottom-center', 'bottom-center'],
-                    ['bottom-top', 'bottom-top'],
+                    [
+                        'label' => '',
+                        'value' => '',
+                    ],
+                    [
+                        'label' => 'top-bottom',
+                        'value' => 'top-bottom',
+                    ],
+                    [
+                        'label' => 'top-center',
+                        'value' => 'top-center',
+                    ],
+                    [
+                        'label' => 'top-top',
+                        'value' => 'top-top',
+                    ],
+                    [
+                        'label' => 'center-bottom',
+                        'value' => 'center-bottom',
+                    ],
+                    [
+                        'label' => 'center-center',
+                        'value' => 'center-center',
+                    ],
+                    [
+                        'label' => 'center-top',
+                        'value' => 'center-top',
+                    ],
+                    [
+                        'label' => 'bottom-bottom',
+                        'value' => 'bottom-bottom',
+                    ],
+                    [
+                        'label' => 'bottom-center',
+                        'value' => 'bottom-center',
+                    ],
+                    [
+                        'label' => 'bottom-top',
+                        'value' => 'bottom-top',
+                    ],
                 ],
             ],
         ],
@@ -152,27 +299,90 @@ defined('TYPO3') || die();
                 'type' => 'select',
                 'renderType' => 'selectSingle',
                 'items' => [
-                    ['', ''],
-                    ['linear', 'linear'],
-                    ['ease', 'ease'],
-                    ['ease-in', 'ease-in'],
-                    ['ease-out', 'ease-out'],
-                    ['ease-in-out', 'ease-in-out'],
-                    ['ease-in-back', 'ease-in-back'],
-                    ['ease-out-back', 'ease-out-back'],
-                    ['ease-in-out-back', 'ease-in-out-back'],
-                    ['ease-in-sine', 'ease-in-sine'],
-                    ['ease-out-sine', 'ease-out-sine'],
-                    ['ease-in-out-sine', 'ease-in-out-sine'],
-                    ['ease-in-quad', 'ease-in-quad'],
-                    ['ease-out-quad', 'ease-out-quad'],
-                    ['ease-in-out-quad', 'ease-in-out-quad'],
-                    ['ease-in-cubic', 'ease-in-cubic'],
-                    ['ease-out-cubic', 'ease-out-cubic'],
-                    ['ease-in-out-cubic', 'ease-in-out-cubic'],
-                    ['ease-in-quart', 'ease-in-quart'],
-                    ['ease-out-quart', 'ease-out-quart'],
-                    ['ease-in-out-quart', 'ease-in-out-quart'],
+                    [
+                        'label' => '',
+                        'value' => '',
+                    ],
+                    [
+                        'label' => 'linear',
+                        'value' => 'linear',
+                    ],
+                    [
+                        'label' => 'ease',
+                        'value' => 'ease',
+                    ],
+                    [
+                        'label' => 'ease-in',
+                        'value' => 'ease-in',
+                    ],
+                    [
+                        'label' => 'ease-out',
+                        'value' => 'ease-out',
+                    ],
+                    [
+                        'label' => 'ease-in-out',
+                        'value' => 'ease-in-out',
+                    ],
+                    [
+                        'label' => 'ease-in-back',
+                        'value' => 'ease-in-back',
+                    ],
+                    [
+                        'label' => 'ease-out-back',
+                        'value' => 'ease-out-back',
+                    ],
+                    [
+                        'label' => 'ease-in-out-back',
+                        'value' => 'ease-in-out-back',
+                    ],
+                    [
+                        'label' => 'ease-in-sine',
+                        'value' => 'ease-in-sine',
+                    ],
+                    [
+                        'label' => 'ease-out-sine',
+                        'value' => 'ease-out-sine',
+                    ],
+                    [
+                        'label' => 'ease-in-out-sine',
+                        'value' => 'ease-in-out-sine',
+                    ],
+                    [
+                        'label' => 'ease-in-quad',
+                        'value' => 'ease-in-quad',
+                    ],
+                    [
+                        'label' => 'ease-out-quad',
+                        'value' => 'ease-out-quad',
+                    ],
+                    [
+                        'label' => 'ease-in-out-quad',
+                        'value' => 'ease-in-out-quad',
+                    ],
+                    [
+                        'label' => 'ease-in-cubic',
+                        'value' => 'ease-in-cubic',
+                    ],
+                    [
+                        'label' => 'ease-out-cubic',
+                        'value' => 'ease-out-cubic',
+                    ],
+                    [
+                        'label' => 'ease-in-out-cubic',
+                        'value' => 'ease-in-out-cubic',
+                    ],
+                    [
+                        'label' => 'ease-in-quart',
+                        'value' => 'ease-in-quart',
+                    ],
+                    [
+                        'label' => 'ease-out-quart',
+                        'value' => 'ease-out-quart',
+                    ],
+                    [
+                        'label' => 'ease-in-out-quart',
+                        'value' => 'ease-in-out-quart',
+                    ],
                 ],
             ],
         ],